+++ src/strings/Counter.java
@@ -5,8 +5,8 @@ import java.util.Map;
 public abstract class Counter implements ICounter{
     static String someString = "foobarbazzzzfufufuf";
 
-    static void print(Map<Character, Integer> charToCount) {
-        for (Map.Entry<Character, Integer> characterToCount : charToCount.entrySet()) {
+    static void print(Map<Character, Long> charToCount) {
+        for (Map.Entry<Character, Long> characterToCount : charToCount.entrySet()) {
             System.out.printf("%s => %s\n", characterToCount.getKey(), characterToCount.getValue());
         }
     }
