+++ src/strings/ICounter.java
@@ -3,10 +3,10 @@ package strings;
 import java.util.Map;
 
 public interface ICounter {
-    static Map<Character, Integer> countChars() {
+    static Map<Character, Long> countChars() {
         return null;
     }
-    static void print(Map<Character, Integer> characterInteger) {
+    static void print(Map<Character, Long> characterInteger) {
 
     }
 }
